+++ Projects/App/Sources/Present/Detail/ViewController/DetailPostViewController.swift
@@ -147,9 +147,15 @@ final class DetailPostViewController: BaseVC<DetailPostViewModel>, CommentDataPr
     }
     
     private func sharePostButtonDidTap() {
+        let backgroundImage = ChoiceAsset.Images.instaBackground.image
+        
         sharePostButton.rx.tap
             .bind(with: self) { owner, _ in
-                owner.shareToInstaStories()
+                owner.detailPostView.contentLabel.numberOfLines = 4
+                ShareToInstagram.shareToInstaStories(detailPostView: owner.detailPostView, backgroundImage: backgroundImage) {
+                    owner.presentFailedShareAlert()
+                }
+                owner.detailPostView.contentLabel.numberOfLines = 0
             }.disposed(by: disposeBag)
     }
     
@@ -438,35 +444,6 @@ final class DetailPostViewController: BaseVC<DetailPostViewModel>, CommentDataPr
                         $0.height.equalTo(newValue.height + 50)
                     }
                 }
-            }
-        }
-    }
-    
-    private func shareToInstaStories() {
-        detailPostView.contentLabel.numberOfLines = 4
-        let renderer = UIGraphicsImageRenderer(bounds: detailPostView.bounds)
-
-        let saveImage = renderer.image { context in
-            detailPostView.layer.render(in: context.cgContext)
-        }
-
-        let appID = "Choice"
-        let backgroundImage = ChoiceAsset.Images.instaBackground.image
-        
-        if let storiesUrl = URL(string: "instagram-stories://share?source_application=\(appID)") {
-            if UIApplication.shared.canOpenURL(storiesUrl) {
-                guard let imageData = saveImage.pngData() else { return }
-                let pasteboardItems: [String: Any] = [
-                    "com.instagram.sharedSticker.stickerImage": imageData,
-                    "com.instagram.sharedSticker.backgroundImage" : backgroundImage
-                ]
-                let pasteboardOptions = [
-                    UIPasteboard.OptionsKey.expirationDate: Date().addingTimeInterval(300)
-                ]
-                UIPasteboard.general.setItems([pasteboardItems], options: pasteboardOptions)
-                UIApplication.shared.open(storiesUrl, options: [:], completionHandler: nil)
-            } else {
-                presentFailedShareAlert()
             }
         }
     }
